+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Change  : Rename deployable_key to component_name for the `glassfish_deployable` resource to bring it inline with
             GlassFish documentation.
 * Bug     : Fix the documentation for the `glassfish_web_env_entry` resource.
+* Bug     : Ensure that the destroy action of the `glassfish_domain` resource runs as the correct user.
 
 ## v0.4.46: